Fixed german MyStuff translations (#37)
* Fixed german name for Blue Toad

*Blau* Toad is non-existant grammar. Did Google Translate do this

* Fixed german name of Chain Chomp

Name has to be hyphenated when shortened, as if not, the words wouldn't share context

* Fixed german name for Dry Bones

sure, the name is Dry Bones but it's a name nontheless, so if shortened, has to be hyphenated and capitalized.

* Fixed german name for Gold Mario

Missing hyphens, yet again

* Fixed kart part names

Gold Mantis had unnecessary space and Fire Swoop was surpassing the string limit, so I shortened it.

* Fixed german name for Mr. L

See https://www.mariowiki.com/Mr._L#Names_in_other_languages

* Update russian name for Midbus

Woah! This time not a fix for german!
See https://www.mariowiki.com/Midbus#Names_in_other_languages
Looks like that wasn't caught.